Make SvgTagTest an integration test.
